Refactor task and user routes, update GoogleTasksService instance creation
Changes made to improve how GoogleTasksService is instantiated and used throughout the application, specifically in task-routes and user-routes. Additional changes were made to seed.js to align with the user list_id changes in task-routes. Also, an error logging feature was added to catch and report issues when getting a list of tasks.